pytest-tap now provides the path to the test
I made so the tap output now provides the path to the tests.  I made this change
to be able to have duplications of testnames but in different parts of a
testsuite. Also this will give better feedback to developers getting the output
whishing to take a look at the test and what it tests in more detail.

closes #28
Signed-off-by: mattemagikern <mansgariusson@gmail.com>