Switch to pytest-cov which has better subprocess support.
The plugin uses subprocess to run tests and the default coverage
configuration was not catching lines that were covered. By switching to
pytest-cov, the coverage report is able to collect all the coverage data
and report the true coverage level (which is good because coverage was
really at 100% instead of the lie of 48%!).